TEDEFO-1415 moving physical model related to code and constants to PhysicalModel.java, removing NoticeSaver.java. Renaming test to SaveNotice...
